refactor(scripts): move merge to /utils
