+++ scripts/utils/index.d.ts
@@ -2,3 +2,4 @@ export { findConfig } from './find-config';
 export { writeConfig } from './write-config';
 export { readConfig } from './read-config';
 export { sh } from './sh';
+export { merge } from './merge';
